+++ apps/main/next.config.js
@@ -34,12 +34,11 @@ const nextConfiguration = {
     // !! WARN !!
     // ignoreBuildErrors: true,
   },
-  // todo: setup redirect when we get rid of {output: 'export'}
-  // redirects: async () => [{
-  //   source: '/',
-  //   destination: '/dex',
-  //   permanent: true
-  // }],
+  redirects: async () => [{
+    source: '/',
+    destination: '/dex',
+    permanent: true
+  }],
 }
 
 const withBundleAnalyzer = require('@next/bundle-analyzer')({
