+++ apps/main/src/app/dao/page.tsx
@@ -1,0 +1,3 @@
+import { App } from './client'
+const Page = () => <App />
+export default Page
