--- apps/main/src/dao/globals.css
@@ -1,1 +1,0 @@
-@import '@ui/styles/base.css';
