--- apps/main/src/dex/globals.css
@@ -1,1 +1,0 @@
-@import '@ui/styles/base.css';
