--- apps/main/src/pages/_app.tsx
@@ -1,8 +1,0 @@
-import '@/globals.css'
-import '@/global-extensions'
-import type { AppProps } from 'next/app'
-import { memo } from 'react'
-
-export default memo(function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-})
