--- apps/main/src/pages/index.tsx
@@ -1,8 +1,0 @@
-import { useEffect } from 'react'
-
-export default function Index() {
-  useEffect(() => {
-    location.href = `/dex${location.search}${location.hash}`
-  }, [])
-  return null
-}
