+++ apps/main/tsconfig.json
@@ -10,14 +10,20 @@
       "@external-rewards": ["../../../packages/external-rewards/src/index.ts"],
       "@curvefi/prices-api": ["../../../packages/prices-api/src/index.ts"],
       "@curvefi/prices-api/*": ["../../../packages/prices-api/src/*"]
-    }
+    },
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
   "exclude": ["node_modules", "cypress"],
   "include": [
-    "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
     "../../packages/curve-ui-kit/**/*.d.ts",
-    "../../packages/ui/**/*.d.ts"
+    "../../packages/ui/**/*.d.ts",
+    "next-env.d.ts",
+    ".next/types/**/*.ts"
   ]
 }
